requirements.txt
Flask
flask-session
msal
python-dotenv
requests